"Underline" errors in the editor
Ace supports "marking" locations in the editor with CSS.
These Markers are not modifications to the editor's content,
but are applied as an overlay at the specified locations.

Since the text isn't affected, text styling CSS is ignored.
Instead, a bottom-border style is applied to give the appearance
of text underlining.

ELM Translator errors are mapped to Ace Marker-like objects and
then passed to the editor component as props.

ANTLR Parser errors (grammar errors) are generated and mapped
to Ace Markers in the component.

To avoid stale Markers, on every update to the error list (fresh
parser results or translator results), all of the
custom Markers are removed, then all new Markers are applied.